bmat: formatting
