+++ Makefile.am
@@ -34,7 +34,7 @@ ACLOCAL_AMFLAGS	= -I m4
 pkginclude_HEADERS =  src/libsemigroups-config.h 
 pkginclude_HEADERS += src/libsemigroups-debug.h 
 pkginclude_HEADERS += src/blocks.h     
-pkginclude_HEADERS += src/bmat.h     
+pkginclude_HEADERS += src/bmat8.h     
 pkginclude_HEADERS += src/cong.h
 pkginclude_HEADERS += src/elements.h   
 pkginclude_HEADERS += src/eltcont.h
@@ -52,7 +52,7 @@ lib_LTLIBRARIES = libsemigroups.la
 
 ## libsemigroups sources
 libsemigroups_la_SOURCES =  src/blocks.cc 
-libsemigroups_la_SOURCES += src/bmat.cc
+libsemigroups_la_SOURCES += src/bmat8.cc
 libsemigroups_la_SOURCES += src/elements.cc  
 libsemigroups_la_SOURCES += src/report.cc 
 libsemigroups_la_SOURCES += src/rws.cc        
@@ -70,7 +70,7 @@ EXTRA_DIST =  tests/catch.hpp
 EXTRA_DIST += README.md LICENSE CPPLINT.cfg .clang-format Doxyfile
 EXTRA_DIST += VERSION
 
-BENCHMARK_LINT_FORMAT =  benchmark/src/bmat.bm.cpp
+BENCHMARK_LINT_FORMAT =  benchmark/src/bmat8.bm.cpp
 BENCHMARK_LINT_FORMAT += benchmark/src/cong.bm.cpp
 BENCHMARK_LINT_FORMAT += benchmark/src/examples.h
 BENCHMARK_LINT_FORMAT += benchmark/src/nridempotents.bm.cpp
@@ -87,7 +87,7 @@ check_PROGRAMS = lstest
 
 lstest_SOURCES =  tests/main.test.cc      
 lstest_SOURCES += tests/blocks.test.cc	  
-lstest_SOURCES += tests/bmat.test.cc
+lstest_SOURCES += tests/bmat8.test.cc
 lstest_SOURCES += tests/cong.test.cc	  
 lstest_SOURCES += tests/elements.test.cc  
 lstest_SOURCES += tests/hpcombi.test.cc
