+++ tests/semigroups.4.test.cc
@@ -16,7 +16,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //
 
-#include "../src/bmat.h"
+#include "../src/bmat8.h"
 #include "../src/semigroups.h"
 #include "catch.hpp"
